+++ app/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+import flask
+
+# from flask_sslify import SSLify
+
+app = flask.Flask(__name__)
+# sslify = SSLify(app)
+
+from app import routes, bot
