+++ app/bot/registration_functions.py
@@ -6,7 +6,7 @@ import json
 import spbu
 import telebot
 
-import functions
+from app.bot import functions
 
 
 def set_next_step(user_id, next_step):
@@ -37,7 +37,7 @@ def get_step(user_id):
 
 
 def select_division(message):
-    from flask_app import bot
+    from app.bot import bot
 
     answer = ""
 
@@ -87,7 +87,8 @@ def select_division(message):
 
 
 def select_study_level(message):
-    from flask_app import bot, start_handler
+    from app.bot import bot
+    from app.bot.message_handlers import start_handler
 
     answer = ""
 
@@ -141,7 +142,7 @@ def select_study_level(message):
 
 
 def select_study_program_combination(message):
-    from flask_app import bot
+    from app.bot import bot
 
     answer = ""
 
@@ -209,7 +210,7 @@ def select_study_program_combination(message):
 
 
 def select_admission_year(message):
-    from flask_app import bot
+    from app.bot import bot
 
     answer = ""
 
@@ -292,7 +293,7 @@ def select_admission_year(message):
 
 
 def select_student_group(message):
-    from flask_app import bot
+    from app.bot import bot
 
     answer = ""
 
@@ -362,8 +363,10 @@ def select_student_group(message):
 
 
 def confirm_choice(message):
-    from flask_app import bot, start_handler, main_keyboard
-    from constants import emoji
+    from app.bot import bot
+    from app.bot.keyboards import main_keyboard
+    from app.bot.message_handlers import start_handler
+    from app.bot.constants import emoji
 
     if message.text == "Все верно":
         sql_con = functions.get_connection()
