+++ app/static/favicon.ico
